fix client/server AES padding issue

+++ Classes/base64.h
@@ -11,7 +11,7 @@
 
 #include <string>
 
-std::string base64_encode(unsigned char const* , size_t len);
+std::string base64_encode(unsigned char const* , unsigned int len);
 std::string base64_decode(std::string const& s);
 
 #endif /* defined(__mcdrunning__base64__) */
